Move action sequence out of detail namespace
